+++ src/plotsp.jl
@@ -1,47 +1,59 @@
 """
     plotsp(X, wl = 1:nco(X); size = (500, 350),
-        color = nothing, nsamp, kwargs...) 
+        color = nothing, nsamp = nothing, kwargs...)
 Plotting spectra.
 * `X` : X-data.
 * `wl` : Column names of `X`. Must be numeric.
-* `color` : Set a unique color (and eventually transparency) to the spectra.
-* 'size' : Resolution (horizontal, vertical) of the figure.
-* `nsamp` : Nb. spectra to plot. If `nothing` (default), all spectra are plotted.
+Keyword arguments:
+* `color` : Set a unique color (and eventually transparency) 
+    to the spectra.
+* 'size' : Size (horizontal, vertical) of the figure.
+* `nsamp` : Nb. spectra (X-rows) to plot. If `nothing`, 
+    all spectra are plotted.
 * `kwargs` : Optional arguments to pass in `Axis` of CairoMakie.
+
 The function plots the rows of `X`.
 
 The user has to specify a backend (e.g. CairoMakie).
 
 ## Examples
 ```julia
-    using JchemoData, JLD2, CairoMakie
-    path_jdat = dirname(dirname(pathof(JchemoData)))
-    db = joinpath(path_jdat, "data/cassav.jld2") 
-    @load db dat
-    pnames(dat)
-    
-    X = dat.X
-    wlst = names(X)
-    wl = parse.(Float64, wlst) 
-    plotsp(X).f
-    plotsp(X; color = (:red, .2)).f
-    plotsp(X, wl; xlabel = "Wavelength (nm)",
-        ylabel = "Absorbance").f
+using JchemoData, JLD2, CairoMakie
+path_jdat = dirname(dirname(pathof(JchemoData)))
+db = joinpath(path_jdat, "data/cassav.jld2") 
+@load db dat
+pnames(dat)
 
-    f, ax = plotsp(X)
-    vlines!(ax, 200)
-    f
+X = dat.X
+wlst = names(X)
+wl = parse.(Float64, wlst) 
+plotsp(X).f
+plotsp(X; color = (:red, .2)).f
+plotsp(X, wl; xlabel = "Wavelength (nm)",
+    ylabel = "Absorbance").f
+
+f, ax = plotsp(X, wl;
+    color = (:red, .2))
+xmeans = colmean(X)
+lines!(ax, wl, xmeans;
+    color = :black, linewidth = 2)
+vlines!(ax, 1200)
+f
 ```
 
 """ 
-function plotsp(X, wl = 1:nco(X); resolution = (500, 350),
+function plotsp(X, wl = 1:nco(X); size = (500, 350),
         color = nothing, nsamp = nothing, kwargs...) 
     X = ensure_mat(X)
     if !isnothing(nsamp)
         X = X[sample(1:nro(X), nsamp; replace = false), :]
     end
-    n, p = size(X)
-    f = Figure(size = resolution)
+    ## For not using function Size
+    ## (conflict with argiment size of Makie)
+    n = nro(X)
+    p = nco(X)
+    ## End
+    f = Figure(size = size)
     ax = Axis(f; kwargs...)
     res = list(Matrix{Float64}, n)
     if isnothing(color)
@@ -64,9 +76,9 @@ function plotsp(X, wl = 1:nco(X); resolution = (500, 350),
         for i = 1:n
             res[i] = hcat([wl ; NaN], [X[i, :] ; NaN])
         end
-        # Same as:
-        # res = [hcat([wl ; NaN], [X[i, :] ; NaN]) for i = 1:n]
-        # End
+        ## Same as:
+        ## res = [hcat([wl ; NaN], [X[i, :] ; NaN]) for i = 1:n]
+        ## End
         res = reduce(vcat, res)
         lines!(ax, res[:, 1], res[:, 2]; color = color)
     end